+++ drivers/plant/device.js
@@ -21,7 +21,7 @@ class plant extends Homey.Device {
             var task = await Homey.ManagerCron.getTask(cronName); 
             this.log("The task exists: " + cronName);
             this.log("task " + JSON.stringify(task));
-            task.on('run', () => this.pollSeverCloud(settings));
+            task.on('run', settings => this.pollSeverCloud(settings));
         } catch (err) {
             if (err.code !== 404) {
                 return this.log(`other cron error: ${err.message}`);
@@ -29,7 +29,7 @@ class plant extends Homey.Device {
             this.log("The task has not been registered yet, registering task: " + cronName);
             try {
                 task = await Homey.ManagerCron.registerTask(cronName, "*/5 * * * *", settings)
-                task.on('run', () => this.pollSeverCloud(settings));
+                task.on('run', settings => this.pollSeverCloud(settings));
             } catch (err) {
                 return this.log(`problem with registering cronjob: ${err.message}`);
             }
@@ -141,8 +141,12 @@ class plant extends Homey.Device {
 
     pollSeverCloud(settings) {
         this.log("pollSeverCloud settings " +  JSON.stringify(settings));
+        // let settings2 = this.getSettings();
+        let name2 = 'zeversolar_plant_' + this.getData().id;
+        this.log("pollSeverCloud device name " +  name2);
+        // this.log("pollSeverCloud device settings " +  JSON.stringify(settings2));
+
         zevercloud.getTodayData(settings).then(data => {
-            let device = this;
             var currentdate =new Date().timeNow();
             this.log("refresh now " + currentdate);