+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.0.6
 
+- fix double device and cron tasks on my side. wait for https://github.com/athombv/homey-apps-sdk-issues/issues/50
+
 ## 0.0.5
 
 - adjust e-total,e-month when unit = MWh